backend/pyproject.toml
[tool.poetry]
name = "owren"
version = "0.1.0"
description = "Just the start of the project see ya soon"
authors = ["Maksat Hemdemow <thehighestintheroom00@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.2"
django-cors-headers = "^3.14.0"
python-decouple = "^3.8"
pillow = "^10.0.0"
django-ckeditor-5 = "^0.2.9"
django-ckeditor = "^6.7.0"
black = "^23.7.0"
djangorestframework = "^3.14.0"
django-jazzmin = "^2.6.0"
psycopg2-binary = "^2.9.7"
python-slugify = "^8.0.1"
flake8 = "^6.1.0"
loguru = "^0.7.0"
newspaper3k = "^0.2.8"
pandas = "^2.0.3"
celery = "^5.3.1"
redis = "^5.0.0"
asyncpg = "^0.28.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
